Ellipse, PixelArray scrolling, and EasingAnimation changes (#23)
* Add ellipse drawing and update EasingAnimation API

Introduces an ellipse drawing function. Refactors EasingAnimation to accept start and end positions as properties instead of constructor arguments. Bumps Python package version to 1.3.1.

BUG FIXES
Text rendering now accounts for camera position. Tile rotation is now stored and exposed in radians instead of degrees.

* Add scroll method and ScrollMode enum to PixelArray

Introduces a scroll(dx, dy, scroll_mode) method to PixelArray, allowing pixel buffer scrolling with configurable edge behavior (SMEAR, ERASE, REPEAT). The ScrollMode enum is defined and exposed to Python.

* Rename event cancellation to unschedule

Renamed the event cancellation function from cancelScheduled to unschedule. Also changed the rotate function argument from 'radians' to 'angle' for consistency in math API.

+++ src/renderer.cpp
@@ -103,13 +103,14 @@ void draw(const Texture& texture, Rect dstRect, const Rect& srcRect)
     dstRect.x -= cameraPos.x;
     dstRect.y -= cameraPos.y;
     const SDL_FRect dstSDLRect = {
-        std::roundf(static_cast<float>(dstRect.x)),
-        std::roundf(static_cast<float>(dstRect.y)),
-        std::roundf(static_cast<float>(dstRect.w)),
-        std::roundf(static_cast<float>(dstRect.h)),
+        static_cast<float>(std::round(dstRect.x)),
+        static_cast<float>(std::round(dstRect.y)),
+        static_cast<float>(std::round(dstRect.w)),
+        static_cast<float>(std::round(dstRect.h)),
     };
     const auto srcSDLRect =
         static_cast<SDL_FRect>(srcRect.w == 0.0 && srcRect.h == 0.0 ? texture.getRect() : srcRect);
+
     SDL_RenderTextureRotated(_renderer, texture.getSDL(), &srcSDLRect, &dstSDLRect,
                              TO_DEGREES(texture.angle), nullptr, flipAxis);
 }
